--- tfbpshiny/modules/analysis_sidebar.py
@@ -1,703 +1,0 @@
-"""Analysis sidebar – module-aware selectors and pairwise comparison controls."""
-
-from __future__ import annotations
-
-from typing import Any
-
-from shiny import module, reactive, render, ui
-
-from tfbpshiny.data_service import (
-    get_dto_config,
-    get_or_create_vdb,
-    get_shared_numeric_columns,
-)
-from tfbpshiny.utils.source_name_lookup import get_source_name_dict
-
-_MODULE_LABELS: dict[str, str] = {
-    "binding": "Binding Analysis",
-    "perturbation": "Perturbation Analysis",
-    "composite": "Comparison Analysis",
-}
-
-
-def _module_type_filter(active_module: str, dataset: dict[str, Any]) -> bool:
-    if not dataset.get("selected"):
-        return False
-    if active_module == "binding":
-        return str(dataset.get("type", "")) == "Binding"
-    if active_module == "perturbation":
-        return str(dataset.get("type", "")) == "Perturbation"
-    if active_module == "composite":
-        return True
-    return False
-
-
-def _standard_sidebar_body() -> ui.Tag:
-    """Sidebar body for binding/perturbation analysis modules."""
-    return ui.div(
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "View Mode"),
-            ui.div(
-                {"class": "group-description"},
-                "Choose how to display the data",
-            ),
-            ui.input_radio_buttons(
-                "view_mode",
-                label=None,
-                choices={
-                    "summary": "Summary",
-                    "correlation": "Correlation",
-                },
-                selected="summary",
-            ),
-        ),
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "Dataset A"),
-            ui.div(
-                {"class": "group-description"},
-                "Primary dataset for analysis",
-            ),
-            ui.input_select(
-                "selected_dataset",
-                label=None,
-                choices={},
-                width="100%",
-            ),
-        ),
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "Pairwise"),
-            ui.div(
-                {"class": "group-description"},
-                "Compare two datasets side by side",
-            ),
-            ui.input_switch(
-                "comparison_mode",
-                label="Comparison mode",
-                value=False,
-            ),
-            ui.input_select(
-                "comparison_dataset",
-                label="Dataset B",
-                choices={},
-                width="100%",
-            ),
-            ui.div(
-                {"style": "display:flex; gap:8px; margin-top:8px;"},
-                ui.input_action_button(
-                    "swap_datasets",
-                    "Swap",
-                    class_="btn btn-sm btn-outline-secondary",
-                ),
-                ui.input_action_button(
-                    "exit_comparison",
-                    "Exit",
-                    class_="btn btn-sm btn-outline-secondary",
-                ),
-            ),
-        ),
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "Value Column"),
-            ui.div(
-                {"class": "group-description"},
-                "Column used for correlation analysis",
-            ),
-            ui.output_ui("value_column_choices"),
-        ),
-    )
-
-
-def _composite_sidebar_static() -> ui.Tag:
-    """Static controls for the composite sidebar (method, filter, logic)."""
-    return ui.div(
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "Comparison Method"),
-            ui.div(
-                {"class": "group-description"},
-                "Statistical method for cross-dataset comparison",
-            ),
-            ui.div(
-                {"class": "dto-radio-row"},
-                ui.input_radio_buttons(
-                    "composite_method",
-                    label=None,
-                    choices={
-                        "dto": "DTO",
-                    },
-                    selected="dto",
-                ),
-                ui.div(
-                    {"class": "dto-links"},
-                    ui.a(
-                        "Paper",
-                        href="https://pubmed.ncbi.nlm.nih.gov/" "32060051/",
-                        target="_blank",
-                        rel="noopener noreferrer",
-                        class_="btn-configure",
-                    ),
-                    ui.a(
-                        "Dataset",
-                        href="https://huggingface.co/datasets/"
-                        "BrentLab/yeast_comparative_analysis",
-                        target="_blank",
-                        rel="noopener noreferrer",
-                        class_="btn-configure",
-                    ),
-                ),
-            ),
-        ),
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "Plot Layout"),
-            ui.div(
-                {"class": "group-description"},
-                "Choose how sources map to plot axes",
-            ),
-            ui.input_radio_buttons(
-                "composite_plot_layout",
-                label=None,
-                choices={
-                    "binding_color": "Color: Binding, Facet: Perturbation",
-                    "perturbation_color": "Color: Perturbation, Facet: Binding",
-                },
-                selected="binding_color",
-            ),
-        ),
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "Binding Sources"),
-            ui.div(
-                {"class": "group-description"},
-                "Select binding datasets to include",
-            ),
-            ui.output_ui("composite_binding_choices"),
-        ),
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "Perturbation Sources"),
-            ui.div(
-                {"class": "group-description"},
-                "Select perturbation datasets to include",
-            ),
-            ui.output_ui("composite_perturbation_choices"),
-        ),
-        ui.div(
-            {"class": "mb-md"},
-            ui.div({"class": "group-header"}, "Filter"),
-            ui.div(
-                {"class": "group-description"},
-                "Set significance thresholds",
-            ),
-            ui.div(
-                {"style": "display:flex; gap:8px; align-items:flex-end;"},
-                ui.div(
-                    {"style": "flex:0 0 auto;"},
-                    ui.input_select(
-                        "composite_filter_operator",
-                        label="Operator",
-                        choices={
-                            "<": "<",
-                            "<=": "<=",
-                            ">": ">",
-                            ">=": ">=",
-                        },
-                        selected=">=",
-                        width="80px",
-                    ),
-                ),
-                ui.div(
-                    {"style": "flex:1;"},
-                    ui.input_numeric(
-                        "composite_filter_threshold",
-                        "Threshold",
-                        value=1.3,
-                        min=0,
-                        max=10.0,
-                        step=0.1,
-                    ),
-                ),
-            ),
-        ),
-    )
-
-
-@module.ui
-def analysis_sidebar_ui() -> ui.Tag:
-    """Render the analysis sidebar panel."""
-    return ui.div(
-        {"class": "context-sidebar", "id": "analysis-sidebar"},
-        ui.div(
-            {"class": "sidebar-header"},
-            ui.output_ui("sidebar_title"),
-            ui.output_ui("comparison_hint"),
-        ),
-        ui.div(
-            {"class": "sidebar-body"},
-            ui.output_ui("sidebar_body_content"),
-        ),
-    )
-
-
-@module.server
-def analysis_sidebar_server(
-    input: Any,
-    output: Any,
-    session: Any,
-    active_module: reactive.Value[str],
-    datasets: reactive.Value[list[dict[str, Any]]],
-    analysis_config: reactive.Value[dict[str, Any]],
-) -> None:
-    """Sync sidebar controls to analysis_config reactive value."""
-
-    def _relevant_datasets() -> list[dict[str, Any]]:
-        current_module = active_module()
-        return [
-            dataset
-            for dataset in datasets()
-            if _module_type_filter(current_module, dataset)
-        ]
-
-    @reactive.calc
-    def _vdb_for_module() -> tuple[Any, list[str]] | None:
-        """
-        Cache VDB and numeric columns for current module's datasets.
-
-        Memoized by the sorted db_names of relevant datasets. This prevents duplicate
-        expensive VDB creation when switching modules.
-
-        """
-        relevant = _relevant_datasets()
-        db_names = [str(d["db_name"]) for d in relevant]
-
-        # Need at least 2 datasets to make correlation meaningful.
-        if len(db_names) < 2:
-            return None
-
-        try:
-            vdb = get_or_create_vdb(db_names)
-            columns = get_shared_numeric_columns(db_names, vdb)
-            return (vdb, columns)
-        except Exception:
-            return None
-
-    def _set_config_if_changed(next_config: dict[str, Any]) -> None:
-        if next_config != analysis_config():
-            analysis_config.set(next_config)
-
-    @render.ui
-    def sidebar_body_content() -> ui.Tag:
-        if active_module() == "composite":
-            return _composite_sidebar_static()
-        return _standard_sidebar_body()
-
-    @render.ui
-    def composite_binding_choices() -> ui.Tag:
-        selected = [d for d in datasets() if d.get("selected")]
-        binding = [d for d in selected if d.get("type") == "Binding"]
-
-        # Filter by DTO links
-        dto_config = get_dto_config()
-        dto_binding_dbs = set(dto_config.get("binding", []))
-
-        if dto_binding_dbs:
-            binding = [d for d in binding if str(d["db_name"]) in dto_binding_dbs]
-
-        if not binding:
-            return ui.p(
-                {"style": "font-size:12px; color:var(--color-text-muted);"},
-                "No binding datasets linked to DTO in active set.",
-            )
-        # Get source name mapping for binding sources
-        source_name_map = get_source_name_dict(datatype="binding")
-        choices = {}
-        for d in binding:
-            db_name = str(d["db_name"])
-            source_key = str(d.get("source_key", ""))
-            # Use source display name if available, otherwise fall back to dataset name
-            display_name = source_name_map.get(source_key, d.get("name", db_name))
-            choices[db_name] = display_name
-        config_selected = analysis_config().get("composite_binding_datasets")
-        if isinstance(config_selected, list) and config_selected:
-            restored = [k for k in config_selected if k in choices]
-            if not restored:
-                restored = list(choices.keys())
-        else:
-            restored = list(choices.keys())
-        return ui.input_checkbox_group(
-            "composite_binding_datasets",
-            label=None,
-            choices=choices,
-            selected=restored,
-        )
-
-    @render.ui
-    def composite_perturbation_choices() -> ui.Tag:
-        selected = [d for d in datasets() if d.get("selected")]
-        perturbation = [d for d in selected if d.get("type") == "Perturbation"]
-
-        # Filter by DTO links
-        dto_config = get_dto_config()
-        dto_perturbation_dbs = set(dto_config.get("perturbation", []))
-
-        if dto_perturbation_dbs:
-            perturbation = [
-                d for d in perturbation if str(d["db_name"]) in dto_perturbation_dbs
-            ]
-
-        if not perturbation:
-            return ui.p(
-                {"style": "font-size:12px; color:var(--color-text-muted);"},
-                "No perturbation datasets linked to DTO in active set.",
-            )
-        # Get source name mapping for perturbation sources
-        source_name_map = get_source_name_dict(datatype="perturbation_response")
-        choices = {}
-        for d in perturbation:
-            db_name = str(d["db_name"])
-            source_key = str(d.get("source_key", ""))
-            # Use source display name if available, otherwise fall back to dataset name
-            display_name = source_name_map.get(source_key, d.get("name", db_name))
-            choices[db_name] = display_name
-        config_selected = analysis_config().get("composite_perturbation_datasets")
-        if isinstance(config_selected, list) and config_selected:
-            restored = [k for k in config_selected if k in choices]
-            if not restored:
-                restored = list(choices.keys())
-        else:
-            restored = list(choices.keys())
-        return ui.input_checkbox_group(
-            "composite_perturbation_datasets",
-            label=None,
-            choices=choices,
-            selected=restored,
-        )
-
-    @render.ui
-    def value_column_choices() -> ui.Tag:
-        relevant = _relevant_datasets()
-        db_names = [str(d["db_name"]) for d in relevant]
-
-        if len(db_names) < 2:
-            return ui.p(
-                {"style": "font-size:12px; color:var(--color-text-muted);"},
-                "Need 2+ datasets for correlation.",
-            )
-
-        # Use cached VDB and columns from _vdb_for_module().
-        cached = _vdb_for_module()
-        if cached is None:
-            return ui.p(
-                {"style": "font-size:12px; color:var(--color-text-muted);"},
-                "Unable to load column metadata. Check connection.",
-            )
-
-        _, columns = cached
-
-        if not columns:
-            return ui.p(
-                {"style": "font-size:12px; color:var(--color-text-muted);"},
-                "No numeric columns found in selected datasets.",
-            )
-
-        choices = {c: c for c in columns}
-
-        # Restore previous selection if still valid.
-        prev = analysis_config().get("correlation_value_column", "")
-        selected = prev if prev in choices else columns[0]
-
-        return ui.input_select(
-            "correlation_value_column",
-            label="Value column",
-            choices=choices,
-            selected=selected,
-            width="100%",
-        )
-
-    @render.ui
-    def sidebar_title() -> ui.Tag:
-        mod = active_module()
-        label = _MODULE_LABELS.get(mod, "Analysis")
-        return ui.div(
-            ui.h2(label),
-            ui.div({"class": "subtitle"}, "Configure your analysis view"),
-        )
-
-    @render.ui
-    def comparison_hint() -> ui.Tag:
-        if active_module() == "selection":
-            return ui.span()
-
-        config = analysis_config()
-        if not config.get("comparison_mode"):
-            return ui.span()
-
-        relevant = _relevant_datasets()
-        names = {
-            str(dataset.get("db_name")): str(dataset.get("name"))
-            for dataset in relevant
-        }
-        db_a = str(config.get("selected_db_name", ""))
-        db_b = str(config.get("comparison_db_name", ""))
-
-        if not db_a or not db_b:
-            return ui.div({"class": "subtitle"}, "Pairwise comparison mode enabled")
-
-        label_a = names.get(db_a, db_a)
-        label_b = names.get(db_b, db_b)
-        return ui.div(
-            {"class": "subtitle"},
-            f"Pairwise: {label_a} vs {label_b}",
-        )
-
-    @reactive.effect
-    def _sync_controls_from_config() -> None:
-        if active_module() in ("selection", "composite"):
-            return
-
-        config = analysis_config()
-        ui.update_radio_buttons(
-            "view_mode", selected=str(config.get("view", "summary"))
-        )
-        ui.update_switch(
-            "comparison_mode", value=bool(config.get("comparison_mode", False))
-        )
-
-    @reactive.effect
-    def _sync_composite_controls_from_config() -> None:
-        if active_module() != "composite":
-            return
-
-        config = analysis_config()
-        ui.update_radio_buttons(
-            "composite_method",
-            selected=str(config.get("composite_method", "dto")),
-        )
-        ui.update_radio_buttons(
-            "composite_plot_layout",
-            selected=str(config.get("composite_plot_layout", "binding_color")),
-        )
-        ui.update_select(
-            "composite_filter_operator",
-            selected=str(config.get("composite_filter_operator", ">=")),
-        )
-        ui.update_numeric(
-            "composite_filter_threshold",
-            value=float(config.get("composite_filter_threshold", 1.3)),
-        )
-
-    @reactive.effect
-    def _update_dataset_choices() -> None:
-        if active_module() in ("selection", "composite"):
-            return
-
-        relevant = _relevant_datasets()
-        choices = {
-            str(dataset["db_name"]): str(dataset["name"]) for dataset in relevant
-        }
-
-        config = analysis_config()
-        selected_db = str(config.get("selected_db_name", ""))
-        comparison_db = str(config.get("comparison_db_name", ""))
-
-        all_keys = list(choices.keys())
-        if not selected_db and all_keys:
-            selected_db = all_keys[0]
-
-        if selected_db not in choices and all_keys:
-            selected_db = all_keys[0]
-
-        if not comparison_db and len(all_keys) >= 2:
-            comparison_db = all_keys[1] if all_keys[1] != selected_db else all_keys[0]
-
-        if comparison_db not in choices and all_keys:
-            comparison_db = all_keys[0]
-
-        ui.update_select(
-            "selected_dataset", choices=choices, selected=selected_db or None
-        )
-        ui.update_select(
-            "comparison_dataset", choices=choices, selected=comparison_db or None
-        )
-
-        next_config = dict(config)
-        if next_config.get("selected_db_name") != selected_db:
-            next_config["selected_db_name"] = selected_db
-
-        if next_config.get("comparison_db_name") != comparison_db:
-            next_config["comparison_db_name"] = comparison_db
-
-        # Compare mode is only meaningful with two datasets.
-        if len(all_keys) < 2 and next_config.get("comparison_mode"):
-            next_config["comparison_mode"] = False
-            if next_config.get("view") == "compare":
-                next_config["view"] = "summary"
-
-        _set_config_if_changed(next_config)
-
-    @reactive.effect
-    def _sync_composite_config() -> None:
-        if active_module() != "composite":
-            return
-
-        current_config = dict(analysis_config())
-
-        try:
-            method = str(input.composite_method())
-            threshold = float(input.composite_filter_threshold())
-            operator = str(input.composite_filter_operator())
-            plot_layout = str(input.composite_plot_layout())
-        except Exception:
-            return
-
-        # Read checkbox selections; fall back to current config if not yet rendered.
-        try:
-            bd_selected = list(input.composite_binding_datasets() or [])
-        except Exception:
-            bd_selected = list(current_config.get("composite_binding_datasets", []))
-
-        try:
-            pr_selected = list(input.composite_perturbation_datasets() or [])
-        except Exception:
-            pr_selected = list(
-                current_config.get("composite_perturbation_datasets", [])
-            )
-
-        next_config = dict(current_config)
-        next_config.update(
-            {
-                "composite_method": method,
-                "composite_filter_threshold": threshold,
-                "composite_filter_operator": operator,
-                "composite_binding_datasets": bd_selected,
-                "composite_perturbation_datasets": pr_selected,
-                "composite_plot_layout": plot_layout,
-            }
-        )
-        _set_config_if_changed(next_config)
-
-    @reactive.effect
-    def _sync_config() -> None:
-        if active_module() == "selection" or active_module() == "composite":
-            return
-
-        current_config = dict(analysis_config())
-
-        # Preserve current config values when inputs aren't available (not mounted yet).
-        # Only update values when inputs successfully return new values.
-
-        # View mode - use current config as base, update only if input is available.
-        view_mode = current_config.get("view", "summary")
-        try:
-            view_mode = str(input.view_mode())
-        except Exception:
-            pass  # Keep current config value
-
-        # Selected dataset - preserve current config if input not available.
-        selected_dataset = current_config.get("selected_db_name", "")
-        try:
-            selected_input = input.selected_dataset()
-            if selected_input:
-                selected_dataset = str(selected_input)
-        except Exception:
-            pass  # Keep current config value
-
-        # Comparison mode - preserve current config if input not available.
-        comparison_mode = current_config.get("comparison_mode", False)
-        try:
-            comparison_mode = bool(input.comparison_mode())
-        except Exception:
-            pass  # Keep current config value
-
-        # Comparison dataset - preserve current config if input not available.
-        comparison_dataset = current_config.get("comparison_db_name", "")
-        try:
-            comparison_input = input.comparison_dataset()
-            if comparison_input:
-                comparison_dataset = str(comparison_input)
-        except Exception:
-            pass  # Keep current config value
-
-        # Value column may not exist if fewer than 2 datasets selected.
-        # Preserve current config value when input not available.
-        value_column = current_config.get("correlation_value_column", "")
-        try:
-            value_column_input = input.correlation_value_column()
-            if value_column_input:
-                value_column = str(value_column_input)
-        except Exception:
-            pass  # Keep current config value
-
-        relevant_db_names = {
-            str(dataset.get("db_name")) for dataset in _relevant_datasets()
-        }
-
-        if selected_dataset not in relevant_db_names and relevant_db_names:
-            selected_dataset = sorted(relevant_db_names)[0]
-
-        if comparison_dataset not in relevant_db_names and relevant_db_names:
-            comparison_dataset = sorted(relevant_db_names)[0]
-
-        if (
-            comparison_mode
-            and selected_dataset
-            and comparison_dataset
-            and selected_dataset == comparison_dataset
-        ):
-            alternatives = [
-                db_name
-                for db_name in sorted(relevant_db_names)
-                if db_name != selected_dataset
-            ]
-            if alternatives:
-                comparison_dataset = alternatives[0]
-            else:
-                comparison_mode = False
-
-        if not comparison_mode and view_mode == "compare":
-            view_mode = "summary"
-
-        next_config = dict(current_config)
-        next_config.update(
-            {
-                "view": view_mode,
-                "selected_db_name": selected_dataset,
-                "comparison_mode": comparison_mode,
-                "comparison_db_name": comparison_dataset,
-                "correlation_value_column": value_column,
-            }
-        )
-        _set_config_if_changed(next_config)
-
-    @reactive.effect
-    @reactive.event(input.swap_datasets, ignore_init=True)
-    def _swap_datasets() -> None:
-        if active_module() == "selection":
-            return
-
-        config = dict(analysis_config())
-        db_a = str(config.get("selected_db_name", ""))
-        db_b = str(config.get("comparison_db_name", ""))
-        if not db_a or not db_b:
-            return
-
-        config["selected_db_name"] = db_b
-        config["comparison_db_name"] = db_a
-        config["comparison_mode"] = True
-        _set_config_if_changed(config)
-
-    @reactive.effect
-    @reactive.event(input.exit_comparison, ignore_init=True)
-    def _exit_comparison() -> None:
-        if active_module() == "selection":
-            return
-
-        config = dict(analysis_config())
-        config["comparison_mode"] = False
-        config["comparison_db_name"] = ""
-        if config.get("view") == "compare":
-            config["view"] = "summary"
-        _set_config_if_changed(config)
